# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/account-groups": {
      "get": {
        "description": "Retrieves all account groups owned by the authenticated user.",
        "operationId": "AccountsController_listGroups",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Account groups retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": [
                      {
                        "id": "f1f3a3e1-5d9f-4584-a704-f0fc641b7788",
                        "name": "Futures Accounts",
                        "description": "Group for derivatives trading accounts"
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "error": {
                      "code": "UNAUTHORIZED",
                      "message": "Authentication is required",
                      "details": []
                    },
                    "meta": {
                      "path": "/api/v1/account-groups",
                      "method": "GET"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List account groups",
        "tags": [
          "Accounts"
        ]
      },
      "post": {
        "description": "Creates a new account group for the authenticated user.",
        "operationId": "AccountsController_createGroup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "The details of the account group to create",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountGroupCreateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account group created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": {
                      "id": "9e305f2e-6f28-4910-a67f-3ba940f6688f",
                      "name": "Scalping Group",
                      "description": "Group for high-frequency setups"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Request payload is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "error": {
                      "code": "VALIDATION_ERROR",
                      "message": "Invalid request payload",
                      "details": [
                        {
                          "field": "name",
                          "message": "Required"
                        }
                      ]
                    },
                    "meta": {
                      "path": "/api/v1/account-groups",
                      "method": "POST"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "error": {
                      "code": "UNAUTHORIZED",
                      "message": "Authentication is required",
                      "details": []
                    },
                    "meta": {
                      "path": "/api/v1/account-groups",
                      "method": "POST"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "An account group with the same name already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "error": {
                      "code": "CONFLICT",
                      "message": "Account group already exists",
                      "details": []
                    },
                    "meta": {
                      "path": "/api/v1/account-groups",
                      "method": "POST"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new account group",
        "tags": [
          "Accounts"
        ]
      }
    },
    "/api/v1/account-groups/{id}": {
      "put": {
        "description": "Updates account group fields by identifier.",
        "operationId": "AccountsController_updateGroup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Account group identifier.",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "example": "f1f3a3e1-5d9f-4584-a704-f0fc641b7788",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Payload to update account group data.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountGroupUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account group updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": {
                      "id": "f1f3a3e1-5d9f-4584-a704-f0fc641b7788",
                      "name": "Updated Futures Accounts",
                      "description": "Updated group description"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Request parameter or payload is invalid."
          },
          "401": {
            "description": "Authentication is required."
          },
          "404": {
            "description": "Account group was not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update account group",
        "tags": [
          "Accounts"
        ]
      }
    },
    "/api/v1/accounts/account-types": {
      "get": {
        "description": "Retrieves static account type select options.",
        "operationId": "AccountsController_listAccountTypeOptions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Account type options retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectOptionListResponseDto_Output"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List account type options",
        "tags": [
          "Accounts"
        ]
      }
    },
    "/api/v1/accounts/currencies": {
      "get": {
        "description": "Retrieves static base currency select options.",
        "operationId": "AccountsController_listCurrencyOptions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Currency options retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectOptionListResponseDto_Output"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List currency options",
        "tags": [
          "Accounts"
        ]
      }
    },
    "/api/v1/accounts/brokers": {
      "get": {
        "description": "Retrieves static broker select options.",
        "operationId": "AccountsController_listBrokerOptions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Broker options retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectOptionListResponseDto_Output"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List broker options",
        "tags": [
          "Accounts"
        ]
      }
    },
    "/api/v1/accounts/timezones": {
      "get": {
        "description": "Retrieves static IANA timezone select options.",
        "operationId": "AccountsController_listTimezoneOptions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Timezone options retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectOptionListResponseDto_Output"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List timezone options",
        "tags": [
          "Accounts"
        ]
      }
    },
    "/api/v1/accounts": {
      "post": {
        "description": "Creates a new trading account for the authenticated user.",
        "operationId": "AccountsController_createAccount",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Payload to create a trading account.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountCreateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": {
                      "id": "5a8f198f-31ef-4584-b806-e4f57ff52cb6",
                      "name": "Binance Futures",
                      "broker": "Binance",
                      "accountType": "CRYPTO"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Request payload is invalid."
          },
          "401": {
            "description": "Authentication is required."
          },
          "409": {
            "description": "An account with the same name already exists."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create account",
        "tags": [
          "Accounts"
        ]
      },
      "get": {
        "description": "Retrieves accounts using optional filtering parameters.",
        "operationId": "AccountsController_listAccounts",
        "parameters": [
          {
            "name": "group_id",
            "required": false,
            "in": "query",
            "description": "Filter accounts by account group identifier.",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "example": "f1f3a3e1-5d9f-4584-a704-f0fc641b7788",
              "type": "string"
            }
          },
          {
            "name": "archived",
            "required": false,
            "in": "query",
            "description": "Filter accounts by archived status.",
            "schema": {
              "enum": [
                "true",
                "false"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Accounts retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": [
                      {
                        "id": "5a8f198f-31ef-4584-b806-e4f57ff52cb6",
                        "name": "Binance Futures",
                        "archivedAt": null
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Query parameters are invalid."
          },
          "401": {
            "description": "Authentication is required."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List accounts",
        "tags": [
          "Accounts"
        ]
      }
    },
    "/api/v1/accounts/{id}": {
      "put": {
        "description": "Updates account details by identifier.",
        "operationId": "AccountsController_updateAccount",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Account identifier.",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "example": "5a8f198f-31ef-4584-b806-e4f57ff52cb6",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Payload to update account fields.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": {
                      "id": "5a8f198f-31ef-4584-b806-e4f57ff52cb6",
                      "name": "Bybit Futures"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Request parameter or payload is invalid."
          },
          "401": {
            "description": "Authentication is required."
          },
          "404": {
            "description": "Account was not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update account",
        "tags": [
          "Accounts"
        ]
      }
    },
    "/api/v1/accounts/{id}/archive": {
      "post": {
        "description": "Archives an account by identifier.",
        "operationId": "AccountsController_archiveAccount",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Account identifier.",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "example": "5a8f198f-31ef-4584-b806-e4f57ff52cb6",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Account archived successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": {
                      "id": "5a8f198f-31ef-4584-b806-e4f57ff52cb6",
                      "archivedAt": "2026-04-05T10:15:30.000Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Path parameter is invalid."
          },
          "401": {
            "description": "Authentication is required."
          },
          "404": {
            "description": "Account was not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Archive account",
        "tags": [
          "Accounts"
        ]
      }
    },
    "/api/v1/accounts/{id}/restore": {
      "post": {
        "description": "Restores an archived account by identifier.",
        "operationId": "AccountsController_restoreAccount",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Account identifier.",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "example": "5a8f198f-31ef-4584-b806-e4f57ff52cb6",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Account restored successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": {
                      "id": "5a8f198f-31ef-4584-b806-e4f57ff52cb6",
                      "archivedAt": null
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Path parameter is invalid."
          },
          "401": {
            "description": "Authentication is required."
          },
          "404": {
            "description": "Account was not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Restore account",
        "tags": [
          "Accounts"
        ]
      }
    },
...
  },
  "info": {
    "title": "CryJournal API",
    "description": "OpenAPI documentation for CryJournal",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AccountGroupCreateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "description": "Account group name (minimum 1 character)."
          },
          "description": {
            "description": "Optional account group description.",
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "AccountGroupUpdateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "description": "Account group name (minimum 1 character)."
          },
          "description": {
            "description": "Optional account group description.",
            "type": "string"
          }
        }
      },
      "SelectOptionListResponseDto_Output": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "value": {
                  "type": "string",
                  "description": "Option value stored in the account payload."
                },
                "label": {
                  "type": "string",
                  "description": "Human-readable option label for UI selects."
                }
              },
              "required": [
                "value",
                "label"
              ],
              "additionalProperties": false
            },
            "description": "List of select options."
          }
        },
        "required": [
          "data"
        ],
        "additionalProperties": false
      },
...
    }
  }
}